Add an elevator entity (BROKEN)
Also test if having the entity move itself with the standing entity
works better or not.

There is some bug where gravity is causing entities to fall through the
platform when it is moving up. The movement code for entities on
platforms is really shitty right now anyway, so we should just rework it
completely.